fix(domain): the offset is one byte, as the original declares it
The previous commit stored the offset in a `uint16_t` and capped it at 255.
The storage is now a `uint8_t`, so the cap is the type. The original
firmware declares `uint8_t offset` (Gravity.ino:69) and PRD 11.1 gives the
field one byte, so this is the faithful behaviour.

The limitation stays as it is: above SUBDIV /2 a step lasts more than 256
ticks, and the offset no longer reaches the end of the step. This is not a
defect to improve.

`setOffset` holds the cap. `clampOffset` then holds only the step limit:
with one byte of storage its own cap became unreachable, so the dead branch
is gone.

Measured: RAM -6 B (six channels, one byte each), Flash -190 B against the
previous commit. The footprint is 1725 / 28774 B, so 410 B stay under the
guard. Peak stack unchanged at 207 B.

+++ include/flexseq/SequencerEngine.h
@@ -74,9 +74,9 @@ public:
     ChannelMode getChannelMode(uint8_t channel) const;
     bool setChannelMode(uint8_t channel, ChannelMode mode);
 
-    static constexpr uint16_t MAX_OFFSET = 255;
+    static constexpr uint8_t MAX_OFFSET = 255;
 
-    uint16_t getOffset(uint8_t channel) const;
+    uint8_t getOffset(uint8_t channel) const;
     bool setOffset(uint8_t channel, uint16_t offset);
 
     uint8_t getSkipChance(uint8_t channel) const;
@@ -118,7 +118,7 @@ private:
         uint16_t ticksPerStep;
         uint8_t barLength;  // graphical measure separation, in steps
         uint8_t mode;
-        uint16_t offset;
+        uint8_t offset;
         uint8_t skipChance;
         uint8_t localStep;  // in [0, effectiveLength)
         uint16_t acc;       // ticks into the current step, in [0, stepTicks)
